[WEB-23] Fix logout issue
There were some properties being set on the User model at the same time browser
data (localStorage data and a cookie) were being deleted. This caused a weird
race condition where the localStorage data would be deleted, but updating the
properties on the User model would cause the data to be serialized back to
localStorage again.

Setting the properties *after* deleting the browser data seems to solve the
issue.