Login should depend on both cookie and localStorage
Users that log in via the old website will get a cookie. Having the cookie
causes the application to assume that the user has already logged in and skips
checking for the localStorage user data. This causes login to fail.

Login now requires the existence of both the cookie and the localStorage user
data. This allows a user that has previously logged in via the old website to
still login to the new website.