Add toJSON data massaging for all models
Most of our models utilize the loading/loaded properties to manage fetch state.
This update forces all models to remove those attributes before saving. This
will also be useful down the road if there are other props we need for the
clientside app but we want to remove before saving.