[WEB-40] Fix profile in Safari and Firefox
Safari and Firefox haven't been doing a great job of respecting the way
Marionette wants to render things. For now I jammed some rerendering based
on collection updates into the view. I'd like to revisit this later since I'm
sure this won't be the only time we encounter the issue.